// src/app/models/user.ts
export class User{
    first_name: string = '';
    last_name: string = '';
    address: string = '';
    city: string = '';
    country: string = '';
    state: string = '';
    zipcode: string = '';
    email: string = '';
    mobile_number: string = '';
    tag: string = '';
    description: string = '';
    password: string = '';
    user_otp: string = '';
}